refactor: split tiles.go (4786 lines) into focused sub-files
Extract four domain-specific files from the tiles.go god file:

- tiles_gauge.go (221 lines): gauge construction and shimmer rendering
- tiles_header.go (646 lines): header/meta/reset/cycle pill rendering,
  Gemini quota logic
- tiles_metrics.go (513 lines): compact metric summaries, metric line
  rendering, formatting helpers
- tiles_composition.go (2684 lines): model/client/vendor/tool/language/
  project composition analysis, bar rendering, color palettes

The remaining tiles.go (746 lines) retains grid layout, tile rendering
dispatch, and loading state.
